move HauserService to internal package
